refactor(ai): move openabcode provider from coding-agent to @openabcode/ai
