+++ packages/ai/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added built-in `openabcode` provider with hosted model catalog.
+
 ## [1.0.6] - 2026-07-16
 
 ## [1.0.5] - 2026-07-16
