+++ packages/ai/src/providers/all.ts
@@ -23,6 +23,7 @@ import { mistralProvider } from "./mistral.ts";
 import { moonshotaiProvider } from "./moonshotai.ts";
 import { moonshotaiCnProvider } from "./moonshotai-cn.ts";
 import { nvidiaProvider } from "./nvidia.ts";
+import { openabcodeProvider } from "./openabcode.ts";
 import { openaiProvider } from "./openai.ts";
 import { openaiCodexProvider } from "./openai-codex.ts";
 import { opencodeProvider } from "./opencode.ts";
@@ -90,6 +91,7 @@ export function builtinProviders(): Provider[] {
 		moonshotaiProvider(),
 		moonshotaiCnProvider(),
 		nvidiaProvider(),
+		openabcodeProvider(),
 		openaiProvider(),
 		openaiCodexProvider(),
 		opencodeProvider(),
