+++ packages/coding-agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Moved OpenABCode provider registration from coding-agent to `@openabcode/ai` as a built-in provider.
+
 ## [1.0.6] - 2026-07-16
 
 ### Changed
