+++ packages/coding-agent/test/suite/router.test.ts
@@ -1,7 +1,8 @@
 import type { Api, Model } from "@openabcode/ai";
 import { fauxAssistantMessage, registerFauxProvider } from "@openabcode/ai/compat";
+import { OPENABCODE_PROVIDER_ID as OPENABCODE_PROVIDER } from "@openabcode/ai/providers/openabcode";
+import { OPENABCODE_HOSTED_UPSTREAM } from "@openabcode/ai/providers/openabcode.models";
 import { describe, expect, it } from "vitest";
-import { OPENABCODE_HOSTED_UPSTREAM, OPENABCODE_PROVIDER } from "../../src/core/openabcode-provider.ts";
 import { classifyProvider, classifyProviderHeuristic, pickRouteModel, routeProviderOf } from "../../src/core/router.ts";
 import { BUILTIN_SLASH_COMMANDS } from "../../src/core/slash-commands.ts";
